# ats-be-transaction-service/dist/generated/swagger.json
{
    "components": {
...
        "schemas": {
...
            "TransactionReportResponse": {
                "properties": {
                    "count": {
                        "type": "number",
                        "format": "double"
                    },
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/TransactionResponse"
                        },
                        "type": "array",
                        "nullable": true
                    }
                },
                "required": [
                    "count",
                    "items"
                ],
                "type": "object"
            },
...
            "ResponseGetReceivingCountries": {
                "properties": {
                    "countryList": {
                        "items": {
                            "$ref": "#/components/schemas/CountryType"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "countryList"
                ],
                "type": "object"
            },
            "ResponseGetTransferType": {
                "properties": {
                    "types": {
                        "items": {
                            "type": "number",
                            "format": "double"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "types"
                ],
                "type": "object"
            },
...
            "ResponseCalculateAmount": {
                "properties": {
                    "amountToCollect": {
                        "type": "number",
                        "format": "double"
                    },
                    "receivingAmount": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "amountToCollect",
                    "receivingAmount"
                ],
                "type": "object"
            },
...
            "ResponseCreateTransaction": {
                "properties": {
                    "message": {
                        "type": "string"
                    },
                    "transaction": {
                        "$ref": "#/components/schemas/TransactionResponse"
                    }
                },
                "required": [
                    "message",
                    "transaction"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "jwt": {
                "type": "jwt",
                "name": "x-access-token",
                "in": "headers"
            },
            "api": {
                "type": "api",
                "name": "x-access-token",
                "in": "headers"
            },
            "api_jwt": {
                "type": "api_jwt",
                "name": "x-access-token",
                "in": "headers"
            }
        }
    },
    "info": {
        "title": "transactions-service",
        "version": "0.1.0",
        "contact": {}
    },
    "openapi": "3.0.0",
    "paths": {
        "/api/transactions": {
            "get": {
                "operationId": "GetTransactions",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionReportResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "api_jwt": [
                            "TRANSACTION_REPORTS"
                        ]
                    }
                ],
                "parameters": []
            }
        },
        "/api/transactions/rate": {
            "get": {
                "operationId": "GetRate",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "number",
                                    "format": "double",
                                    "nullable": true
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/transactions/fee": {
            "get": {
                "operationId": "GetFee",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "number",
                                    "format": "double",
                                    "nullable": true
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/transactions/receivingCountries": {
            "get": {
                "operationId": "GetReceivingCountries",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseGetReceivingCountries"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": []
            }
        },
        "/api/transactions/{countryId}/transferTypes": {
            "get": {
                "operationId": "GetTransactionTypesForCountry",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseGetTransferType"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "countryId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/transactions/calculateAmount": {
            "post": {
                "operationId": "CalculateAmount",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseCalculateAmount"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestCalculateAmount"
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/create": {
            "post": {
                "operationId": "CreateTransaction",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ResponseCreateTransaction"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": [
                            "SEND_MONEY"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "header",
                        "name": "user-id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "header",
                        "name": "user-type",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/UserType"
                        }
                    },
                    {
                        "in": "header",
                        "name": "user-agent-id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "header",
                        "name": "user-branch-id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestCreateTransaction"
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/{countryId}/banks": {
            "get": {
                "operationId": "GetBanksByCountry",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/Bank"
                                    },
                                    "type": "array",
                                    "nullable": true
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "countryId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/transactions/cancel-request": {
            "post": {
                "operationId": "CreateTransactionCancelRequest",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionCrResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": [
                            "SEND_MONEY"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionCancelRequest"
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/change-name-request": {
            "post": {
                "operationId": "CreateTransactionChangeNameRequest",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionCrResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": [
                            "SEND_MONEY"
                        ]
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionChangeNameRequest"
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/allowed-statuses/{status}": {
            "get": {
                "operationId": "GetApplicableStatuses",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/AllowedStatusesResponse"
                                        }
                                    ],
                                    "nullable": true
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "status",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/transactions/{id}/update": {
            "put": {
                "operationId": "Update",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/TransactionResponse"
                                        }
                                    ],
                                    "nullable": true
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": [
                            "SEND_MONEY"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "header",
                        "name": "user-id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RequestUpdateTransaction"
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/check-compliance": {
            "post": {
                "operationId": "CheckComplianceRules",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ComplianceCheckRequest"
                            }
                        }
                    }
                }
            }
        },
        "/api/transactions/update-status/{transactionId}/{status}": {
            "put": {
                "operationId": "UpdateStatus",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "transactionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "status",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/TransactionStatus"
                        }
                    }
                ]
            }
        },
        "/api/transactions/reports-count/{type}": {
            "get": {
                "operationId": "GetTransactionCount",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "number",
                                    "format": "double"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": [
                            "TRANSACTION_REPORTS"
                        ]
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "type",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/TransactionCountType"
                        }
                    }
                ]
            }
        },
        "/api/transactions/country/{id}": {
            "get": {
                "operationId": "GetCountryById",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountryResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [
                    {
                        "jwt": []
                    }
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
...
        "/api/internal/transactions/country/{id}": {
            "get": {
                "operationId": "GetCountryById",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CountryResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ]
            }
        },
        "/api/internal/transactions/update-status/{transactionId}/{status}": {
            "put": {
                "operationId": "UpdateStatus",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "transactionId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "in": "path",
                        "name": "status",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/TransactionStatus"
                        }
                    }
                ]
            }
        },
        "/api/internal/transactions/reports-count/{type}": {
            "get": {
                "operationId": "GetTransactionCount",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "number",
                                    "format": "double"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Transaction"
                ],
                "security": [],
                "parameters": [
                    {
                        "in": "path",
                        "name": "type",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/TransactionCountType"
                        }
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}